fix: some fixes

+++ components/FollowButton.tsx
@@ -12,45 +12,45 @@ interface FollowButtonProps {
 export default function FollowButton({ userId, initialState }: FollowButtonProps) {
 
     const { data } = useFollowersInfo(userId, initialState)
-
-    console.log(typeof data.followers);
-
     const queryClient = useQueryClient()
-    const { mutate } = useMutation({
-        mutationFn: () => data.isFollowedByUser ?
-            fetch(`/api/users/${userId}/followers`, {
-                method: "DELETE"
-            }) :
-            fetch(`/api/users/${userId}/followers`, {
-                method: "POST"
-            }),
-        onSuccess: async (res) => {
+    const { mutate, isPending } = useMutation({
+        mutationFn: async ({ isFollowing }: { isFollowing: boolean }) => {
+            if (isFollowing) {
+
+                const res = await fetch(`/api/users/${userId}/followers`, {
+                    method: "DELETE"
+                })
+                if (!res.ok) throw new Error("Failed")
+                return res.json()
+            } else {
+
+                const res = await fetch(`/api/users/${userId}/followers`, {
+                    method: "POST"
+                })
+                if (!res.ok) throw new Error("Failed")
+                return res.json()
+            }
+        },
+        onMutate: async ({ isFollowing }) => {
             const queryKey: QueryKey = ["follower-info", userId]
             await queryClient.cancelQueries({ queryKey })
-
             const prevState = queryClient.getQueryData<FollowerInfo>(queryKey)
-
-            console.log("from::", prevState?.followers);
-            console.log("to::", (prevState?.followers || 0) + (prevState?.isFollowedByUser ? -1 : 1));
-
-
-            queryClient.setQueryData<FollowerInfo>(queryKey, () => (
-                {
-                    followers: (prevState?.followers || 0) + (prevState?.isFollowedByUser ? -1 : 1),
-                    isFollowedByUser: !prevState?.isFollowedByUser
-                }))
-
+            if (!prevState) return prevState;
+            queryClient.setQueryData<FollowerInfo>(queryKey, () => ({
+                followers: (prevState?.followers || 0) + (isFollowing ? -1 : 1),
+                isFollowedByUser: !prevState?.isFollowedByUser
+            }))
             return { prevState }
         },
         onError(error, variables, context) {
             const queryKey: QueryKey = ["follower-info", userId]
             queryClient.setQueryData(queryKey, context?.prevState)
         },
-
     })
     return (
         <Button
-            onClick={() => mutate()}
+            disabled={isPending}
+            onClick={() => mutate({ isFollowing: data.isFollowedByUser })}
             variant={data.isFollowedByUser ? "secondary" : "default"}
         >
             {data.isFollowedByUser ? "Unfollow" : "Follow"}
